pyproject.toml
[tool.poetry]
name = "gea"
version = "0.2.50"
description = "Gestión de Expedientes de Agrimensores"
authors = ["Santiago Pestarini <santiagonob@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/quijot/gea-package"
packages = [{ include = "gea" }]

[tool.poetry.dependencies]
python = ">=3.6,<4"
django = "^3.1.5"
django-nested-admin = "^3.3.0"
django-extensions = "^2.2.9"
django-crispy-forms = "^1.11.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
black = "^19.10b0"
ipython = "^7.13.0"
pydot = "^1.4.1"
psycopg2 = "^2.8.5"
dj_database_url = "^0.5.0"
gunicorn = "^20.0.4"
whitenoise = "^5.0.1"

[tool.black]
line-length = 88
target-version = ["py36"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
